refactor: move util folder to root repository and configured to reflect this change everywhere.
