+++ learn-to-bank/app/components/UpcomingPaymentsList.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { Text } from '@geist-ui/core';
-import { GeneratedTransaction, getTransactionsForPeriodBulk } from '~/util/futureTransactionUtil';
+import { GeneratedTransaction, getTransactionsForPeriodBulk } from '@parent/learn-to-bank-util/futureTransactionUtil';
 import { addMonths, startOfDay } from 'date-fns';
 import { RecurringTransactionWithRecipient } from '~/routes/app.upcoming';
